refactor(suite-native): extract useCloseDeviceCompromisedScreen
